Moved adagrams.rb into lib folder
